+++ hypothesis-python/src/hypothesis/core.py
@@ -385,9 +385,7 @@ def get_random_for_wrapped_test(test, wrapped_test):
         return Random(seed)
 
 
-def process_arguments_to_given(
-    wrapped_test, arguments, kwargs, given_kwargs, argspec,
-):
+def process_arguments_to_given(wrapped_test, arguments, kwargs, given_kwargs, argspec):
     selfy = None
     arguments, kwargs = convert_positional_arguments(wrapped_test, arguments, kwargs)
 
@@ -479,7 +477,7 @@ def new_given_argspec(original_argspec, given_kwargs):
 
 class StateForActualGivenExecution:
     def __init__(
-        self, test_runner, search_strategy, test, settings, random, wrapped_test,
+        self, test_runner, search_strategy, test, settings, random, wrapped_test
     ):
         self.test_runner = test_runner
         self.search_strategy = search_strategy
@@ -1027,7 +1025,7 @@ def given(
                 )
 
             state = StateForActualGivenExecution(
-                test_runner, search_strategy, test, settings, random, wrapped_test,
+                test_runner, search_strategy, test, settings, random, wrapped_test
             )
 
             reproduce_failure = wrapped_test._hypothesis_internal_use_reproduce_failure
@@ -1164,12 +1162,12 @@ def given(
             )
             random = get_random_for_wrapped_test(test, wrapped_test)
             _args, _kwargs, test_runner, search_strategy = process_arguments_to_given(
-                wrapped_test, (), {}, given_kwargs, argspec,
+                wrapped_test, (), {}, given_kwargs, argspec
             )
             assert not _args
             assert not _kwargs
             state = StateForActualGivenExecution(
-                test_runner, search_strategy, test, settings, random, wrapped_test,
+                test_runner, search_strategy, test, settings, random, wrapped_test
             )
             digest = function_digest(test)
 
@@ -1225,9 +1223,9 @@ def find(
     specifier: SearchStrategy[Ex],
     condition: Callable[[Any], bool],
     *,
-    settings: Settings = None,
-    random: Random = None,
-    database_key: bytes = None
+    settings: Optional[Settings] = None,
+    random: Optional[Random] = None,
+    database_key: Optional[bytes] = None
 ) -> Ex:
     """Returns the minimal example from the given strategy ``specifier`` that
     matches the predicate function ``condition``."""