+++ hypothesis-python/src/hypothesis/internal/charmap.py
@@ -52,10 +52,7 @@ def charmap():
         f = charmap_file()
         try:
             with gzip.GzipFile(f, "rb") as i:
-                # When the minimum Python 3 version becomes 3.6, this can be
-                # simplified to `json.load(i)` without needing to decode first.
-                data = i.read().decode()
-                tmp_charmap = dict(json.loads(data))
+                tmp_charmap = dict(json.load(i))
 
         except Exception:
             # This loop is reduced to using only local variables for performance;