+++ hypothesis-python/src/hypothesis/strategies/_internal/core.py
@@ -298,12 +298,12 @@ def none() -> SearchStrategy[None]:
 
 @overload
 def one_of(args: Sequence[SearchStrategy[Any]]) -> SearchStrategy[Any]:
-    pass  # pragma: no cover
+    raise NotImplementedError
 
 
 @overload  # noqa: F811
 def one_of(*args: SearchStrategy[Any]) -> SearchStrategy[Any]:
-    pass  # pragma: no cover
+    raise NotImplementedError
 
 
 def one_of(*args):  # noqa: F811
@@ -344,7 +344,9 @@ def one_of(*args):  # noqa: F811
 
 @cacheable
 @defines_strategy(force_reusable_values=True)
-def integers(min_value: int = None, max_value: int = None) -> SearchStrategy[int]:
+def integers(
+    min_value: Optional[int] = None, max_value: Optional[int] = None,
+) -> SearchStrategy[int]:
     """Returns a strategy which generates integers.
 
     If min_value is not None then all values will be >= min_value. If
@@ -414,11 +416,11 @@ def booleans() -> SearchStrategy[bool]:
 @defines_strategy(force_reusable_values=True)
 @deprecated_posargs
 def floats(
-    min_value: Real = None,
-    max_value: Real = None,
+    min_value: Optional[Real] = None,
+    max_value: Optional[Real] = None,
     *,
-    allow_nan: bool = None,
-    allow_infinity: bool = None,
+    allow_nan: Optional[bool] = None,
+    allow_infinity: Optional[bool] = None,
     width: int = 64,
     exclude_min: bool = False,
     exclude_max: bool = False
@@ -441,8 +443,6 @@ def floats(
     required to represent the generated float. Valid values are 16, 32, or 64.
     Passing ``width=32`` will still use the builtin 64-bit ``float`` class,
     but always for values which can be exactly represented as a 32-bit float.
-    Half-precision floats (``width=16``) are not supported on Python 3.5,
-    unless :pypi:`Numpy` is installed.
 
     The exclude_min and exclude_max argument can be used to generate numbers
     from open or half-open intervals, by excluding the respective endpoints.
@@ -472,10 +472,6 @@ def floats(
         raise InvalidArgument(
             "Got width=%r, but the only valid values are the integers 16, "
             "32, and 64." % (width,)
-        )
-    if width == 16 and sys.version_info[:2] < (3, 6) and "numpy" not in sys.modules:
-        raise InvalidArgument(  # pragma: no cover
-            "width=16 requires either Numpy, or Python >= 3.6"
         )
 
     check_valid_bound(min_value, "min_value")
@@ -647,13 +643,13 @@ def tuples(*args: SearchStrategy) -> SearchStrategy[tuple]:
 
 @overload
 def sampled_from(elements: Sequence[T]) -> SearchStrategy[T]:
-    pass  # pragma: no cover
+    raise NotImplementedError
 
 
 @overload  # noqa: F811
 def sampled_from(elements: Type[enum.Enum]) -> SearchStrategy[Any]:
     # `SearchStrategy[Enum]` is unreliable due to metaclass issues.
-    pass  # pragma: no cover
+    raise NotImplementedError
 
 
 @defines_strategy(try_non_lazy=True)  # noqa: F811
@@ -681,7 +677,7 @@ def sampled_from(elements):
         repr_ = "sampled_from(%s.%s)" % (elements.__module__, elements.__name__)
     else:
         repr_ = "sampled_from(%r)" % (elements,)
-    if hasattr(enum, "Flag") and isclass(elements) and issubclass(elements, enum.Flag):
+    if isclass(elements) and issubclass(elements, enum.Flag):
         # Combinations of enum.Flag members are also members.  We generate
         # these dynamically, because static allocation takes O(2^n) memory.
         # LazyStrategy is used for the ease of force_repr.
@@ -699,8 +695,8 @@ def lists(
     elements: SearchStrategy[Ex],
     *,
     min_size: int = 0,
-    max_size: int = None,
-    unique_by: UniqueBy = None,
+    max_size: Optional[int] = None,
+    unique_by: Optional[UniqueBy] = None,
     unique: bool = False
 ) -> SearchStrategy[List[Ex]]:
     """Returns a list containing values drawn from elements with length in the
@@ -787,7 +783,7 @@ def lists(
 @defines_strategy()
 @deprecated_posargs
 def sets(
-    elements: SearchStrategy[Ex], *, min_size: int = 0, max_size: int = None
+    elements: SearchStrategy[Ex], *, min_size: int = 0, max_size: Optional[int] = None,
 ) -> SearchStrategy[Set[Ex]]:
     """This has the same behaviour as lists, but returns sets instead.
 
@@ -807,7 +803,7 @@ def sets(
 @defines_strategy()
 @deprecated_posargs
 def frozensets(
-    elements: SearchStrategy[Ex], *, min_size: int = 0, max_size: int = None
+    elements: SearchStrategy[Ex], *, min_size: int = 0, max_size: Optional[int] = None,
 ) -> SearchStrategy[FrozenSet[Ex]]:
     """This is identical to the sets function but instead returns
     frozensets."""
@@ -837,8 +833,8 @@ def iterables(
     elements: SearchStrategy[Ex],
     *,
     min_size: int = 0,
-    max_size: int = None,
-    unique_by: UniqueBy = None,
+    max_size: Optional[int] = None,
+    unique_by: Optional[UniqueBy] = None,
     unique: bool = False
 ) -> SearchStrategy[Iterable[Ex]]:
     """This has the same behaviour as lists, but returns iterables instead.
@@ -861,7 +857,7 @@ def iterables(
 def fixed_dictionaries(
     mapping: Dict[T, SearchStrategy[Ex]],
     *,
-    optional: Dict[T, SearchStrategy[Ex]] = None
+    optional: Optional[Dict[T, SearchStrategy[Ex]]] = None
 ) -> SearchStrategy[Dict[T, Ex]]:
     """Generates a dictionary of the same type as mapping with a fixed set of
     keys mapping to strategies. ``mapping`` must be a dict subclass.
@@ -907,7 +903,7 @@ def dictionaries(
     *,
     dict_class: type = dict,
     min_size: int = 0,
-    max_size: int = None
+    max_size: Optional[int] = None
 ) -> SearchStrategy[Dict[Ex, T]]:
     # Describing the exact dict_class to Mypy drops the key and value types,
     # so we report Dict[K, V] instead of Mapping[Any, Any] for now.  Sorry!
@@ -939,12 +935,12 @@ def dictionaries(
 @deprecated_posargs
 def characters(
     *,
-    whitelist_categories: Sequence[str] = None,
-    blacklist_categories: Sequence[str] = None,
-    blacklist_characters: Sequence[str] = None,
-    min_codepoint: int = None,
-    max_codepoint: int = None,
-    whitelist_characters: Sequence[str] = None
+    whitelist_categories: Optional[Sequence[str]] = None,
+    blacklist_categories: Optional[Sequence[str]] = None,
+    blacklist_characters: Optional[Sequence[str]] = None,
+    min_codepoint: Optional[int] = None,
+    max_codepoint: Optional[int] = None,
+    whitelist_characters: Optional[Sequence[str]] = None
 ) -> SearchStrategy[str]:
     r"""Generates characters, length-one :class:`python:str`\ ings,
     following specified filtering rules.
@@ -1050,7 +1046,7 @@ def text(
     ),
     *,
     min_size: int = 0,
-    max_size: int = None
+    max_size: Optional[int] = None
 ) -> SearchStrategy[str]:
     """Generates strings with characters drawn from ``alphabet``, which should
     be a collection of length one strings or a strategy generating such strings.
@@ -1141,7 +1137,9 @@ def from_regex(
 @cacheable
 @defines_strategy(force_reusable_values=True)
 @deprecated_posargs
-def binary(*, min_size: int = 0, max_size: int = None) -> SearchStrategy[bytes]:
+def binary(
+    *, min_size: int = 0, max_size: Optional[int] = None,
+) -> SearchStrategy[bytes]:
     """Generates :class:`python:bytes`.
 
     The generated :class:`python:bytes` will have a length of at least ``min_size``
@@ -1161,7 +1159,7 @@ def binary(*, min_size: int = 0, max_size: int = None) -> SearchStrategy[bytes]:
 @cacheable
 @defines_strategy()
 def randoms(
-    *, note_method_calls: bool = False, use_true_random: bool = None
+    *, note_method_calls: bool = False, use_true_random: Optional[bool] = None,
 ) -> SearchStrategy[random.Random]:
     """Generates instances of ``random.Random``. The generated Random instances
     are of a special HypothesisRandom subclass.
@@ -1447,13 +1445,6 @@ def _from_type(thing: Type[Ex]) -> SearchStrategy[Ex]:
         if getattr(thing, "__origin__", None) is typing.Union:
             args = sorted(thing.__args__, key=types.type_sorting_key)
             return one_of([from_type(t) for t in args])
-    # We can't resolve forward references, and under Python 3.5 (only)
-    # a forward reference is an instance of type.  Hence, explicit check:
-    elif type(thing) == getattr(typing, "_ForwardRef", None):  # pragma: no cover
-        raise ResolutionFailed(
-            "thing=%s cannot be resolved.  Upgrading to python>=3.6 may "
-            "fix this problem via improvements to the typing module." % (thing,)
-        )
     if not types.is_a_type(thing):
         raise InvalidArgument("thing=%s must be a type" % (thing,))
     # Now that we know `thing` is a type, the first step is to check for an
@@ -1522,10 +1513,10 @@ def _from_type(thing: Type[Ex]) -> SearchStrategy[Ex]:
 @defines_strategy(force_reusable_values=True)
 @deprecated_posargs
 def fractions(
-    min_value: Union[Real, str] = None,
-    max_value: Union[Real, str] = None,
+    min_value: Optional[Union[Real, str]] = None,
+    max_value: Optional[Union[Real, str]] = None,
     *,
-    max_denominator: int = None
+    max_denominator: Optional[int] = None
 ) -> SearchStrategy[Fraction]:
     """Returns a strategy which generates Fractions.
 
@@ -1606,7 +1597,7 @@ def fractions(
 
 
 def _as_finite_decimal(
-    value: Union[Real, str, None], name: str, allow_infinity: Optional[bool],
+    value: Union[Real, str, None], name: str, allow_infinity: Optional[bool]
 ) -> Optional[Decimal]:
     """Convert decimal bounds to decimals, carefully."""
     assert name in ("min_value", "max_value")
@@ -1632,12 +1623,12 @@ def _as_finite_decimal(
 @defines_strategy(force_reusable_values=True)
 @deprecated_posargs
 def decimals(
-    min_value: Union[Real, str] = None,
-    max_value: Union[Real, str] = None,
+    min_value: Optional[Union[Real, str]] = None,
+    max_value: Optional[Union[Real, str]] = None,
     *,
-    allow_nan: bool = None,
-    allow_infinity: bool = None,
-    places: int = None
+    allow_nan: Optional[bool] = None,
+    allow_infinity: Optional[bool] = None,
+    places: Optional[int] = None
 ) -> SearchStrategy[Decimal]:
     """Generates instances of :class:`python:decimal.Decimal`, which may be:
 
@@ -1837,9 +1828,9 @@ def composite(f: Callable[..., Ex]) -> Callable[..., SearchStrategy[Ex]]:
 def complex_numbers(
     *,
     min_magnitude: Real = 0,
-    max_magnitude: Real = None,
-    allow_infinity: bool = None,
-    allow_nan: bool = None
+    max_magnitude: Optional[Real] = None,
+    allow_infinity: Optional[bool] = None,
+    allow_nan: Optional[bool] = None
 ) -> SearchStrategy[complex]:
     """Returns a strategy that generates complex numbers.
 
@@ -1920,7 +1911,9 @@ def complex_numbers(
 
 
 @deprecated_posargs
-def shared(base: SearchStrategy[Ex], *, key: Hashable = None) -> SearchStrategy[Ex]:
+def shared(
+    base: SearchStrategy[Ex], *, key: Optional[Hashable] = None,
+) -> SearchStrategy[Ex]:
     """Returns a strategy that draws a single shared value per run, drawn from
     base. Any two shared instances with the same key will share the same value,
     otherwise the identity of this strategy will be used. That is:
@@ -1943,7 +1936,7 @@ def shared(base: SearchStrategy[Ex], *, key: Hashable = None) -> SearchStrategy[
 @cacheable
 @defines_strategy(force_reusable_values=True)
 @deprecated_posargs
-def uuids(*, version: int = None) -> SearchStrategy[UUID]:
+def uuids(*, version: Optional[int] = None) -> SearchStrategy[UUID]:
     """Returns a strategy that generates :class:`UUIDs <uuid.UUID>`.
 
     If the optional version argument is given, value is passed through
@@ -2192,7 +2185,7 @@ def emails() -> SearchStrategy[str]:
 def functions(
     *,
     like: Callable[..., Any] = lambda: None,
-    returns: SearchStrategy[Any] = None,
+    returns: Optional[SearchStrategy[Any]] = None,
     pure: bool = False
 ) -> SearchStrategy[Callable[..., Any]]:
     # The proper type signature of `functions()` would have T instead of Any, but mypy