+++ hypothesis-python/src/hypothesis/strategies/_internal/random.py
@@ -142,7 +142,7 @@ def define_copy_method(name):
 
     spec = inspect.getfullargspec(STUBS.get(name, target))
 
-    result = define_function_signature(target.__name__, target.__doc__, spec,)(
+    result = define_function_signature(target.__name__, target.__doc__, spec)(
         implementation
     )
 
@@ -326,8 +326,7 @@ class ArtificialRandom(HypothesisRandom):
             result = self.__data.draw(st.floats(min_value=0.0))
         elif method == "shuffle":
             result = self.__data.draw(st.permutations(range(len(kwargs["x"]))))
-        # This is tested for but only appears in 3.9 so doesn't appear in
-        # coverage.
+        # This is tested for but only appears in 3.9 so doesn't appear in coverage.
         elif method == "randbytes":  # pragma: no cover
             n = kwargs["n"]
             result = self.__data.draw(st.binary(min_size=n, max_size=n))