+++ hypothesis-python/src/hypothesis/strategies/_internal/types.py
@@ -73,9 +73,8 @@ def try_issubclass(thing, superclass):
     superclass = getattr(superclass, "__origin__", None) or superclass
     try:
         return issubclass(thing, superclass)
-    except (AttributeError, TypeError):  # pragma: no cover
-        # Some types can't be the subject or object of an instance or
-        # subclass check under Python 3.5
+    except (AttributeError, TypeError):
+        # Some types can't be the subject or object of an instance or subclass check
         return False
 
 
@@ -85,10 +84,8 @@ def is_a_new_type(thing):
     return (
         hasattr(thing, "__supertype__")
         and (
-            hasattr(typing, "NewType")
-            and getattr(thing, "__module__", None) == "typing"
-            or hasattr(typing_extensions, "NewType")
-            and getattr(thing, "__module__", None) == "typing_extensions"
+            getattr(thing, "__module__", None) == "typing"
+            or getattr(thing, "__module__", None) == "typing_extensions"
         )
         and inspect.isfunction(thing)
     )
@@ -148,9 +145,8 @@ def _try_import_forward_ref(thing, bound):  # pragma: no cover
             and getattr(thing, "__module__", None) == "typing"
         ):
             raise ResolutionFailed(
-                "Looks like you are using python3.6 or any of the previous versions "
-                "together with a TypeVar bound to a ForwardRef. "
-                "It is not supported. Upgrading to python3.7+ will solve this problem."
+                "It looks like you're using a TypeVar bound to a ForwardRef on Python "
+                "3.6, which is not supported - try ugrading to Python 3.7 or later."
             ) from None
         # We fallback to `ForwardRef` instance, you can register it as a type as well:
         # >>> from typing import ForwardRef
@@ -168,15 +164,6 @@ def from_typing_type(thing):
     # information to sensibly resolve to strategies at runtime.
     # Finally, we run a variation of the subclass lookup in `st.from_type`
     # among generic types in the lookup.
-    #
-    # Under 3.6 Union is handled directly in st.from_type, as the argument is
-    # not an instance of `type`. However, under Python 3.5 Union *is* a type
-    # and we have to handle it here, including failing if it has no parameters.
-    if hasattr(thing, "__union_params__"):  # pragma: no cover
-        args = sorted(thing.__union_params__ or (), key=type_sorting_key)
-        if not args:
-            raise ResolutionFailed("Cannot resolve Union of no types.")
-        return st.one_of([st.from_type(t) for t in args])
     if getattr(thing, "__origin__", None) == tuple or isinstance(
         thing, getattr(typing, "TupleMeta", ())
     ):
@@ -229,13 +216,12 @@ def from_typing_type(thing):
         if is_generic_type(k) and try_issubclass(k, thing)
     }
     if typing.Dict in mapping:
-        # The subtype relationships between generic and concrete View types
-        # are sometimes inconsistent under Python 3.5, so we pop them out to
-        # preserve our invariant that all examples of from_type(T) are
-        # instances of type T - and simplify the strategy for abstract types
-        # such as Container
-        for t in (typing.KeysView, typing.ValuesView, typing.ItemsView):
-            mapping.pop(t, None)
+        # ItemsView can cause test_lookup.py::test_specialised_collection_types
+        # to fail, due to weird isinstance behaviour around the elements.
+        mapping.pop(typing.ItemsView, None)
+        if sys.version_info[:2] == (3, 6):  # pragma: no branch
+            # `isinstance(dict().values(), Container) is False` on py36 only -_-
+            mapping.pop(typing.ValuesView, None)
     if len(mapping) > 1:
         # issubclass treats bytestring as a kind of sequence, which it is,
         # but treating it as such breaks everything else when it is presumed
@@ -249,8 +235,6 @@ def from_typing_type(thing):
         elem_type = (getattr(thing, "__args__", None) or ["not int"])[0]
         if getattr(elem_type, "__origin__", None) is typing.Union:
             union_elems = elem_type.__args__
-        elif hasattr(elem_type, "__union_params__"):  # pragma: no cover
-            union_elems = elem_type.__union_params__  # python 3.5 only
         else:
             union_elems = ()
         if not any(
@@ -436,10 +420,8 @@ _global_type_lookup.update(
             # As with Reversible, we tuplize this for compatibility with Hashable.
             st.lists(st.integers(0, 255)).map(tuple),  # type: ignore
         ),
-        # xIO are only available in .io on Python 3.5, but available directly
-        # as typing.*IO from 3.6 onwards and mypy 0.730 errors on the compat form.
-        typing.io.BinaryIO: st.builds(io.BytesIO, st.binary()),  # type: ignore
-        typing.io.TextIO: st.builds(io.StringIO, st.text()),  # type: ignore
+        typing.BinaryIO: st.builds(io.BytesIO, st.binary()),
+        typing.TextIO: st.builds(io.StringIO, st.text()),
     }
 )
 if hasattr(typing, "SupportsIndex"):  # pragma: no cover  # new in Python 3.8
@@ -471,7 +453,7 @@ def register(type_, fallback=None, *, module=typing):
     return inner
 
 
-@register("Type")
+@register(typing.Type)
 @register("Type", module=typing_extensions)
 def resolve_Type(thing):
     if getattr(thing, "__args__", None) is None:
@@ -479,17 +461,14 @@ def resolve_Type(thing):
     args = (thing.__args__[0],)
     if getattr(args[0], "__origin__", None) is typing.Union:
         args = args[0].__args__
-    elif hasattr(args[0], "__union_params__"):  # pragma: no cover
-        args = args[0].__union_params__
-    if isinstance(ForwardRef, type):  # pragma: no cover
-        # Duplicate check from from_type here - only paying when needed.
-        for a in args:
-            if type(a) == ForwardRef:
-                raise ResolutionFailed(
-                    "thing=%s cannot be resolved.  Upgrading to "
-                    "python>=3.6 may fix this problem via improvements "
-                    "to the typing module." % (thing,)
-                )
+    # Duplicate check from from_type here - only paying when needed.
+    for a in args:
+        if type(a) == ForwardRef:
+            raise ResolutionFailed(
+                "thing=%s cannot be resolved.  Upgrading to "
+                "python>=3.6 may fix this problem via improvements "
+                "to the typing module." % (thing,)
+            )
     return st.sampled_from(sorted(args, key=type_sorting_key))
 
 
@@ -541,7 +520,7 @@ def resolve_Dict(thing):
     )
 
 
-@register("DefaultDict", st.builds(collections.defaultdict))
+@register(typing.DefaultDict, st.builds(collections.defaultdict))
 @register("DefaultDict", st.builds(collections.defaultdict), module=typing_extensions)
 def resolve_DefaultDict(thing):
     return resolve_Dict(thing).map(lambda d: collections.defaultdict(None, d))