+++ hypothesis-python/tests/conjecture/test_engine.py
@@ -1272,7 +1272,7 @@ def test_database_contains_only_pareto_front():
         runner = ConjectureRunner(
             test,
             settings=settings(
-                max_examples=500, database=db, suppress_health_check=HealthCheck.all(),
+                max_examples=500, database=db, suppress_health_check=HealthCheck.all()
             ),
             database_key=b"stuff",
         )